#4 Fix: 자기 자신 차단 오류 방지

+++ src/main/java/com/apps/pochak/block/service/BlockService.java
@@ -3,6 +3,7 @@ package com.apps.pochak.block.service;
 import com.apps.pochak.block.domain.Block;
 import com.apps.pochak.block.domain.repository.BlockRepository;
 import com.apps.pochak.follow.domain.repository.FollowRepository;
+import com.apps.pochak.global.api_payload.exception.GeneralException;
 import com.apps.pochak.like.domain.repository.LikeRepository;
 import com.apps.pochak.login.jwt.JwtService;
 import com.apps.pochak.member.domain.Member;
@@ -11,6 +12,8 @@ import com.apps.pochak.post.domain.repository.PostRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+
+import static com.apps.pochak.global.api_payload.code.status.ErrorStatus.BLOCK_ONESELF;
 
 @Service
 @RequiredArgsConstructor
@@ -27,6 +30,11 @@ public class BlockService {
     public void blockMember(String handle) {
         Member blocker = jwtService.getLoginMember();
         Member blockedMember = memberRepository.findByHandle(handle, blocker);
+
+        if (blocker.getId().equals(blockedMember.getId())) {
+            throw new GeneralException(BLOCK_ONESELF);
+        }
+
         saveBlockEntity(blockedMember, blocker);
         followRepository.deleteFollowsBetweenMembers(blockedMember, blocker);
         likeRepository.deleteLikesBetweenMembers(blockedMember, blocker);
